+++ lang/cs/fields.php
@@ -1,11 +1,13 @@
 <?php
 
 return [
-    'name' => 'Jméno',
-    'email' => 'E-mail',
-    'email_contact' => 'Zanechte svůj e-mail pro kontakt',
-    'message' => 'Popište svůj problém',
-    'phone' => 'Telefon',
-    'optional' => 'nepovinné',
+
+    'email' => 'Vaše e-mailová adresa',
+    'email_optional' => 'Zanechte svůj e-mail pro kontakt',
+    'message' => 'Sem napište svou zprávu',
+    'name' => 'Vaše celé jméno',
+    'optional' => 'volitelné',
+    'phone' => 'Vaše telefonní číslo',
     'submit' => 'Odeslat',
+
 ];