+++ resources/views/components/link.blade.php
@@ -1,0 +1,10 @@
+@props([
+    'formId',
+])
+
+<a
+    {{ $attributes }}
+    href="Javascript:void(0)"
+    x-data
+    x-on:click.prevent="Livewire.dispatch('openFeedback', {id: '{{ $formId }}'})"
+>{{ $slot ?? $formId }}</a>